building_search that outputs building info
Scrapped RPI building information/maps website and obtained the description of each building if the website provides it and the lat and long of each building. Using the lat and long and google maps link, the program takes it and inserts it into google maps link and returns the link for the directions to each building. Using the description and directions to each building, it compares the arg the user enters on discord and only outputs the information for that one building requested or multiple that have the same word in it if the full building name isn't entered.