SQL-based rewrite of course search
After lots of experimenting with different approaches to a new implementation of course search, we've set up a MySQL database to hold data from our catalog scraper, which is now used in this iteration of the course search. Instead of a request being made on every command, it simply does a regex search on the database contents, which is far more reliable and offers more room for search improvements.