Create calendar menu prototype
After lots of experimenting with the Discord UI kit and the interactions part of the discord.py library, a first iteration of a navigable academic calendar menu is complete.

Right now, it works off data from a local JSON, but the plan is to have the menu request data from our SQL database on each interaction.

src/assets/acadcal_2024.json will soon be removed from this repository.