Add timeout function, course not offered message
Added a bunch of code to get timeout functionality working. The interactions can't stay around forever, so this will communicate to the user when the menus are no longer active. They timeout after 2 minutes of inactivity, or 5 minutes total.
In addition, a message was added to replace "Credits" showing None, which states that the course was not offered during the displayed semester.